# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample service for managing drivers space",
        "title": "DriversSpace API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Selmi hamza",
            "email": "hamza@craftfoundry.tech"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "api-company.herokuapp.com",
    "basePath": "/",
    "paths": {
...
        "/vehicle": {
            "post": {
                "description": "Create a new vehicle with the input payload",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "driver_cars"
                ],
                "summary": "Create a new vehicle",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "enter driver id (tester avec l'id : 11469 ) : ",
                        "name": "DriverID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "enter driver company id (tester avec l'id : 10725 ) : ",
                        "name": "DriverCompanyID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "enter saas office id (tester avec l'id : 1 ) : ",
                        "name": "SaasOfficeID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "enter Color:",
                        "name": "Color",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "enter Matricule:",
                        "name": "Matricule",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "enter Year:",
                        "name": "Year",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "enter Model:",
                        "name": "Model",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "enter Mark:",
                        "name": "Mark",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "enter assur file:",
                        "name": "ASSUR",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "enter cgv file:",
                        "name": "CGV",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "enter atr file:",
                        "name": "ATR",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Models.DriverCar"
                        }
                    }
                }
            }
        },
        "/vehicle_categories/{saas_officeID}": {
            "get": {
                "description": "retrieve a Vehicule categories in database with the saas office ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicle_categorie"
                ],
                "summary": "retrieve a Vehicule categories by saas office ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "enter saas_officeID:",
                        "name": "saas_officeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Models.VehicleCategorie"
                        }
                    }
                }
            }
        },
        "/vehicle_options/{saas_officeID}": {
            "get": {
                "description": "retrieve a Vehicule options in database with the office ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicle_options"
                ],
                "summary": "retrieve a Vehicule options by saas office ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "enter saas_officeID:",
                        "name": "saas_officeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Models.VehicleOption"
                        }
                    }
                }
            }
        },
        "/vehicle_types/{saas_officeID}": {
            "get": {
                "description": "retrieve a Vehicule types in database with the saas office ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicle_types"
                ],
                "summary": "retrieve a Vehicule types by saas office ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "enter saas_officeID:",
                        "name": "saas_officeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Models.VehicleType"
                        }
                    }
                }
            }
        },
        "/vehicle_zones/{saas_officeID}": {
            "get": {
                "description": "retrieve a list of zones  in database by saas office id in request parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "zones"
                ],
                "summary": "retrieve a list of zones by saas office id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "enter saas_office ID:",
                        "name": "saas_officeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Models.Zone"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Models.DriverCar": {
            "type": "object",
            "properties": {
                "carMarque": {
                    "type": "string"
                },
                "carModel": {
                    "type": "string"
                },
                "carStatus": {
                    "type": "integer"
                },
                "carType": {
                    "type": "integer"
                },
                "color": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deleted": {
                    "type": "boolean"
                },
                "driverCompanyID": {
                    "type": "integer"
                },
                "driverID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isFr": {
                    "type": "boolean"
                },
                "matricule": {
                    "type": "string"
                },
                "refusalReason": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "validated": {
                    "type": "integer"
                },
                "vehicleTypeID": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "Models.VehicleCategorie": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "saasOfficeID": {
                    "type": "integer"
                },
                "totalCapacity": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "Models.VehicleOption": {
            "type": "object",
            "properties": {
                "alwaysAvailable": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "deleted": {
                    "type": "boolean"
                },
                "driverOption": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "isDfaultForVehicleOptionCategory": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "priceImpact": {
                    "type": "integer"
                },
                "saasOfficeID": {
                    "type": "integer"
                },
                "timeImpact": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "vehicleOptionCategoryID": {
                    "type": "integer"
                }
            }
        },
        "Models.VehicleType": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "designation": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isAvailable": {
                    "type": "boolean"
                },
                "kmPrice": {
                    "type": "integer"
                },
                "logoVehicle": {
                    "type": "string"
                },
                "maxPlace": {
                    "type": "integer"
                },
                "minPrice": {
                    "type": "integer"
                },
                "pickupPrice": {
                    "type": "integer"
                },
                "saasCompanyID": {
                    "type": "integer"
                },
                "saasOfficeID": {
                    "type": "integer"
                },
                "typeCode": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "Models.Zone": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deleted": {
                    "type": "boolean"
                },
                "driverZone": {
                    "type": "boolean"
                },
                "greenZone": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "latMax": {
                    "type": "number"
                },
                "latMin": {
                    "type": "number"
                },
                "longMax": {
                    "type": "number"
                },
                "longMin": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "saasCompanyID": {
                    "type": "integer"
                },
                "saasOfficeID": {
                    "type": "integer"
                },
                "serviceArea": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}